% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{keepSamples}
\alias{keepSamples}
\title{keepSamples}
\usage{
keepSamples(pcr, keep)
}
\arguments{
\item{pcr}{qPCR objects to work on}

\item{keep}{character vector of samples to keep}
}
\description{
keep only selected samples
}
\details{
details
}

