% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{keepTargets}
\alias{keepTargets}
\title{keepTargets}
\usage{
keepTargets(pcr, keep)
}
\arguments{
\item{pcr}{qPCR objects to work on}

\item{keep}{character vector of targets to keep}
}
\description{
keep only selected targets
}
\details{
details
}

