% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mergePCR}
\alias{mergePCR}
\title{mergePCR}
\usage{
mergePCR(pcrs, ref_sample, ref_gene)
}
\arguments{
\item{ref_sample}{sample to normalize to}

\item{pcr1, pcr2}{qPCR objects to work on}
}
\description{
merge 2 qPCR objects
}
\details{
details
}

