% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotRawCt}
\alias{plotRawCt}
\title{---------------------------------------------------------------------------- #
plotRawCt}
\usage{
plotRawCt(pcr, raw = TRUE)
}
\arguments{
\item{pcr}{qPCR object to work on}
}
\description{
plot raw Ct values for reference
}
\details{
details
}
