% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workingClasses.R
\docType{class}
\name{qPCR-class}
\alias{qPCR-class}
\title{An S4 class to represent a qPCR experiment.}
\description{
An S4 class to represent a qPCR experiment.
}
\section{Slots}{

\describe{
\item{\code{raw.data}}{data.table with raw qPCR data}

\item{\code{metadata}}{data.table with experimental metadata}

\item{\code{design}}{factor vector of sample groupings for barplot coloring}

\item{\code{data}}{processed data (e.g. recalculated relative expression)}
}}

