% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{qPCRsummary}
\alias{qPCRsummary}
\title{qPCRsummary}
\usage{
qPCRsummary(...)
}
\arguments{
\item{LS}{a list of qPCR objects to work on}

\item{keep}{character vector of samples to keep}
}
\description{
make a summary table of qPCR list
}
\details{
details
}
