% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{relExp}
\alias{relExp}
\title{relExp}
\usage{
relExp(pcr, ref_sample, ref_target)
}
\arguments{
\item{pcr}{qPCR object to work on}

\item{groups}{character vector of sample groups}
}
\description{
calculate relative expression for a qPCR run
}
\details{
details
}

