% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{renameSamples}
\alias{renameSamples}
\title{renameSamples}
\usage{
renameSamples(pcr, ...)
}
\arguments{
\item{pcr}{qPCR object to work on}

\item{groups}{character vector of sample groups}
}
\description{
set new sample names for a qPCR object
}

