% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{renameTargets}
\alias{renameTargets}
\title{renameTargets}
\usage{
renameTargets(pcr, old, new)
}
\arguments{
\item{pcr}{qPCR object to work on}

\item{old}{character vector of target names to be replaced}

\item{new}{character vector of new target names}
}
\description{
set new target names for a qPCR object
}
