% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reorderSamples}
\alias{reorderSamples}
\title{reorderSamples}
\usage{
reorderSamples(pcr, old, new)
}
\arguments{
\item{pcr}{qPCR objects to work on}

\item{old}{old sample order}

\item{new}{new sample order}
}
\description{
set the order of samples
}
\details{
details
}
