% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setDesign}
\alias{setDesign}
\title{setDesign}
\usage{
setDesign(pcr, ...)
}
\arguments{
\item{pcr}{qPCR object to work on}

\item{groups}{character vector of sample groups}
}
\description{
set design of a qPCR experiment
}
\details{
details
}

