% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mergePCR}
\alias{mergePCR}
\title{mergePCR}
\usage{
mergePCR(pcr1, pcr2, ref_sample)
}
\arguments{
\item{pcr1, pcr2}{qPCR objects to work on}

\item{ref_sample}{sample to normalize to}
}
\description{
merge 2 qPCR objects
}
\details{
details
}

