% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mergePCR}
\alias{mergePCR}
\title{mergePCR}
\usage{
mergePCR(pcrs, ref_sample, ref_target)
}
\arguments{
\item{pcrs}{a list of qPCR objects to work on}

\item{ref_sample}{sample to normalize to}
}
\description{
merge 2 qPCR objects
}
\details{
details
}
