% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\docType{methods}
\name{plotRQ}
\alias{plotRQ}
\alias{plotRQ,plotRQ-method}
\alias{plotRQ,qPCR-method}
\title{---------------------------------------------------------------------------- #
plotRQ}
\usage{
plotRQ(pcr, minimal)

\S4method{plotRQ}{qPCR}(pcr)
}
\arguments{
\item{filename}{}
}
\description{
plot relative quantification as returned by the machine
}
\details{
details
}

