% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\docType{methods}
\name{plotRQ}
\alias{plotRQ}
\alias{plotRQ,qPCR-method}
\alias{plotRQ,plotRQ-method}
\title{---------------------------------------------------------------------------- #
plotRQ}
\usage{
plotRQ(pcr, raw = FALSE, minimal = FALSE, plottype = c("bar", "point"),
  reverseGroups = FALSE, ...)

\S4method{plotRQ}{qPCR}(pcr, raw = FALSE, minimal = FALSE,
  plottype = c("bar", "point"), reverseGroups = FALSE, ...)
}
\arguments{
\item{filename}{}
}
\description{
plot relative quantification as returned by the machine
}
\details{
details
}
