% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workingClasses.R
\docType{class}
\name{qPCR-class}
\alias{qPCR-class}
\title{An S4 class to represent a qPCR experiment.}
\description{
An S4 class to represent a qPCR experiment.
}
\section{Slots}{

\describe{
\item{\code{data}}{data.table with qPCR data}

\item{\code{metadata}}{data.table with experimental metadata}
}}

